% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webfonts_functions.R
\name{list}
\alias{list}
\title{Retrieves the list of fonts currently served by the Google Fonts Developer API}
\usage{
list(sort = NULL)
}
\arguments{
\item{sort}{Enables sorting of the list}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item 
}

Set \code{options(googleAuthR.scopes.selected = c()}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/fonts/docs/developer_api}{Google Documentation}
}

