% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webfonts_objects.R
\name{Webfont.files}
\alias{Webfont.files}
\title{Webfont.files Object}
\usage{
Webfont.files()
}
\value{
Webfont.files object
}
\description{
Webfont.files Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The font files (with all supported scripts) for each one of the available variants, as a key : value map.
}
\seealso{
Other Webfont functions: \code{\link{Webfont}}
}
