% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webfonts_objects.R
\name{Webfont}
\alias{Webfont}
\title{Google Fonts Developer API Objects 
Accesses the metadata for all families served by Google Fonts, providing a list of families currently available (including available styles and a list of supported script subsets).}
\usage{
Webfont(Webfont.files = NULL, category = NULL, family = NULL,
  files = NULL, lastModified = NULL, subsets = NULL, variants = NULL,
  version = NULL)
}
\arguments{
\item{Webfont.files}{The \link{Webfont.files} object or list of objects}

\item{category}{The category of the font}

\item{family}{The name of the font}

\item{files}{The font files (with all supported scripts) for each one of the available variants, as a key : value map}

\item{lastModified}{The date (format 'yyyy-MM-dd') the font was modified for the last time}

\item{subsets}{The scripts supported by the font}

\item{variants}{The available variants for the font}

\item{version}{The font version}
}
\value{
Webfont object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:50:29
filename: /Users/mark/dev/R/autoGoogleAPI/googlewebfontsv1.auto/R/webfonts_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Webfont Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Webfont functions: \code{\link{Webfont.files}}
}

