% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webfonts_objects.R
\name{WebfontList}
\alias{WebfontList}
\title{WebfontList Object}
\usage{
WebfontList(items = NULL)
}
\arguments{
\item{items}{The list of fonts currently served by the Google Fonts API}
}
\value{
WebfontList object
}
\description{
WebfontList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
