% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webfonts_functions.R
\docType{package}
\name{webfonts_googleAuthR}
\alias{webfonts_googleAuthR}
\alias{webfonts_googleAuthR-package}
\title{Google Fonts Developer API
Accesses the metadata for all families served by Google Fonts, providing a list of families currently available (including available styles and a list of supported script subsets).}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 20:22:03
filename: /Users/mark/dev/R/autoGoogleAPI/googlewebfontsv1.auto/R/webfonts_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item 
}
}
